requirements.txt
pymupdf
python-dotenv
google-genai
